Support to conditionally enable test by implementation type.
