+++ src/test-integration/java/org/opendcs/regression_tests/LoadingAppTestIT.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.junit.jupiter.api.Test;
 import org.opendcs.fixtures.AppTestBase;
-import org.opendcs.fixtures.EnableIfTsDb;
 import org.opendcs.fixtures.annotations.ConfiguredField;
+import org.opendcs.fixtures.annotations.EnableIfTsDb;
 
 import decodes.tsdb.CompAppInfo;
 import decodes.tsdb.TimeSeriesDb;
